Move IsFIPSDisabledByEnv to lib/utils/aws/awsfips
